+++ XilinxProcessorIPLib/drivers/canfd/src/xcanfd_intr.c
@@ -539,6 +539,6 @@ int XCanFd_SetHandler(XCanFd *InstancePtr, u32 HandlerType,
 		break;
 
 	}
-	return Status;
+	return (s32)Status;
 }
 /** @} */
